✅(jest) add a custom resolver to enforce the use of common js
Since jest 28, module import logic has changed and packages which have
es module exports could encounter problem with jest-environment-jsdom as jest
uses the es module exports instead of the common js. To address this issue we
have to use a custom resolver, to force the use of the common js for
problematic packages (nanoid).

https://jestjs.io/docs/next/upgrading-to-jest28#packagejson-exports