🔧(vite) improve dev experience
After a first usage of the monorepo, it seems the developer experience could be
improved. First we prevent dev processes to be killed as soon as there is a
typescript error within the lib. Furthermore, we update the log level from info
to warn to not pollute the console with useless information. Then in order to
speed up build process during development, we disable minify process. Finally,
we only build lib as es module since umd does not allow to break lib into chunks