🔧(vite) setup resolve.alias to import modules with absolute paths
Use relative paths to import local modules is not convenient for development.
It's not straightforward to find the location of a module and furthermore when
we move a file, it's a chore to update all related paths.
So we decided to setup alias to resolve our local modules. First we have to
update the vite config of our projects by setting resolve.alias