🔧(vite) fix dynamic import warning
vite used @rollup/plugin-dynamic-import-vars under the hood to resolve dynamic
imports. Currently, in our codebase, we use this feature to import lazily
translations. But, as the doc of this plugin mentions, this kind of import must
use relative path and we are currently use an alias which resolve to an absolute
 path. In order to remove the warning, we have to use a relative path instead of
  the aliased path.