🔧(lib) export tests stuff into a dedicated chunk
Lib test stuffs contains factories to ease test writing. In some case,
it could be interesting to be able to access to those factories through
an external context. So we have to include this code within the build.
But we do not want to expose this module in the main codebase. So that's
 why we add a second entry point into the vite config to build two
 chunks one containing the business code and other containing the test
 stuff.